+++ src/Transformers/PrincipalComponentAnalysis.php
@@ -7,6 +7,9 @@ use Rubix\ML\DataType;
 use Rubix\ML\Persistable;
 use Rubix\ML\Datasets\Dataset;
 use Rubix\ML\Traits\AutotrackRevisions;
+use Rubix\ML\Specifications\ExtensionIsLoaded;
+use Rubix\ML\Specifications\SpecificationChain;
+use Rubix\ML\Specifications\ExtensionMinimumVersion;
 use Rubix\ML\Specifications\SamplesAreCompatibleWithTransformer;
 use Rubix\ML\Exceptions\InvalidArgumentException;
 use Rubix\ML\Exceptions\RuntimeException;
@@ -72,6 +75,11 @@ class PrincipalComponentAnalysis implements Transformer, Stateful, Persistable
      */
     public function __construct(int $dimensions)
     {
+        SpecificationChain::with([
+            ExtensionIsLoaded::with('tensor'),
+            ExtensionMinimumVersion::with('tensor', '2.1.4'),
+        ])->check();
+
         if ($dimensions < 1) {
             throw new InvalidArgumentException('Dimensions must be'
                 . " greater than 0, $dimensions given.");